Refactor parser methods to StringUtil
