+++ src/main/java/seedu/duke/logic/parser/DeleteCommandParser.java
@@ -12,9 +12,9 @@ import static seedu.duke.commons.core.CommandFormat.DELETE_TASK_FORMAT;
 import static seedu.duke.commons.core.CommandFormat.DELETE_LESSON_FORMAT;
 import static seedu.duke.commons.core.CommandFormat.DELETE_MODULE_FORMAT;
 import static seedu.duke.commons.core.CommandFormat.promptFormat;
+import static seedu.duke.commons.util.StringUtil.removeFirstParam;
 import static seedu.duke.logic.parser.ParserUtil.parseCommandType;
 import static seedu.duke.logic.parser.ParserUtil.parseToZeroIndex;
-import static seedu.duke.logic.parser.ParserUtil.removeFirstParam;
 
 //@@author richwill28
 public class DeleteCommandParser {
