+++ src/main/java/seedu/duke/logic/parser/DoneCommandParser.java
@@ -8,9 +8,9 @@ import seedu.duke.logic.parser.exceptions.ParseException;
 
 import static seedu.duke.commons.core.CommandFormat.DONE_TASK_FORMAT;
 import static seedu.duke.commons.core.CommandFormat.promptFormat;
+import static seedu.duke.commons.util.StringUtil.removeFirstParam;
 import static seedu.duke.logic.parser.ParserUtil.parseCommandType;
 import static seedu.duke.logic.parser.ParserUtil.parseToZeroIndex;
-import static seedu.duke.logic.parser.ParserUtil.removeFirstParam;
 
 //@@author rebchua39
 public class DoneCommandParser {
